7-Zip Version 4.58
- BUG was fixed: 7-Zip showed incorrect timestamps in ISO files.
- BUG in 4.58 alpha 5/6/7 was fixed:
    LZMA encoder worked incorrectly, if lp != 0.
- Unicode (UTF-8) support for filenames in .ZIP archives. Now there are 3 modes:
    1) Default mode: 7-Zip uses UTF-8, if the local code page doesn't contain required symbols.
    2) -mcu switch:  7-Zip uses UTF-8, if there are non-ASCII symbols.
    3) -mcl switch:  7-Zip uses local code page.
- Now it's possible to use -mSW- and -mSW+ switches instead of -mSW=off and -mSW=on
- BUG was fixed: BZip2Encoder and BZip2Decoder used CEvent objects without
  creating, when BZip2 code was called with one thread (with -mmt1 switch or with
  default switches on single thread CPU).
- .lzma support.
- RPM and NSIS support was improved.
- LZMA now stores only (2 << n) or (3 << n) dictionary size value to LZMA properties.
- NTFS time extra in ZIP.
- New item property - kpidTimeType - VT_UI4 (0 - NTFS, 1 - Unix, 2 - DOS).
- Static CRC table is not required now for Lzma Encoder (in Lz MatchFinder).
- Creation time (-mtc switch) for .7z archives
- LZMA encoder was converted to ANSI-C
- Speed optimizations for LZMA decoding. Now it uses C code instead of C++.
- 7-Zip now has 128 MB dictionary limit for 32-bit version:
  It's for speed optimization: kNumLogBits = 9 + sizeof(size_t) / 2;
- TAR: 'D' link flag support.
- 7-Zip now can unpack multivolume RAR archives created with
  "old style volume names" scheme (-vn switch) and names *.001, *.002, ...
- Fixed bugs:
  - 7-Zip FM could not copy / move files to root network folders like \\COMPNAME\FOLDERNAME\
    In case of move it removed original files.
  - SFX-WIN: if there are errors, it still could return 0.
  - ZIP (.XPS file) isZip64 && thisDiskNumber16 == 0xFFFF.
  - ZIP name updating:
      If zip file contains extra field and you try to change properties of files,
      7-zip tries to delete all extra fileds (except for WzAES).
      And that code could hang.
  -  7-Zip GUI didn't suggest BZip2 dictionary size used in previous run.
  -  If creation time stamp was included in .RAR archive, 7-zip used creation time stamp
     as modification time stamp.
- Small changes in Deflate and LZMA compression.
- Some speed optimizations.

+++ CPP/7zip/Archive/ArchiveExports.cpp
@@ -44,9 +44,9 @@ int FindFormatCalssId(const GUID *clsID)
   if (cls != CLSID_CArchiveHandler)
     return -1;
   Byte id = CLS_ARC_ID_ITEM(*clsID);
-  for (UInt32 i = 0; i < g_NumArcs; i++)
+  for (unsigned i = 0; i < g_NumArcs; i++)
     if (g_Arcs[i]->ClassId == id)
-      return i;
+      return (int)i;
   return -1;
 }
